totalTimeNotWriting fix
This should fix the totalTimeNotWriting. This could probably further be tweaked by making a threshold match the yellow warning, but I think ignore stops of less than 1 second is sufficient to start. Now that secondsSinceLastWord is actually in seconds, and longestStretchNotWriting and totalTimeNotWriting are in seconds, no need to divide by 1000 in for SprintRunStat.