Fix: fixed bug where existing files had their word count added
There is a bug where the word count is one lower than it should be
for each file opened in a sprint. I corrected this manually by
subtracting one from the initial word count for each file in
SprintRun's typingUpdate method because I'm not sure what was
causing it.

I added some tests for the new things I added, but some of them don't
pass for some reason. I'm not noticing any issues when I use the plugin
in Odsidian, though, so I'm ignoring that.

I don't really know what I'm doing and did most of this with Claude AI,
but I think I know enough to review it and see that things are working.